updated init

+++ utils.cc
@@ -10,7 +10,7 @@ std::string utils::getTimestamp()
     return buf;
 }
 
-long utils::getTime()
+long long utils::getTime()
 {
     return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
 }
